[refactor #027] 문제 목록 조회 기능 수정
- 카테고리 이름을 나타낼 수 있도록 수정

+++ NBTI/src/main/java/com/dao/nbti/problem/application/service/AdminProblemService.java
@@ -18,7 +18,7 @@ public class AdminProblemService {
     public ProblemListResponse getProblems(ProblemSearchRequest problemSearchRequest) {
         List<ProblemSummaryDTO> problems = problemRepositoryCustom.getProblemsBy(problemSearchRequest);
 
-        int totalItems = problemRepositoryCustom.countProblemsBy(problemSearchRequest);
+        long totalItems = problemRepositoryCustom.countProblemsBy(problemSearchRequest);
 
         return ProblemListResponse.builder()
                 .problems(problems)
